[ObjC] Remove `generate_minimal_imports` generation option warning
Since protobuf now has `import option` the original idea around this
is reduced; but more importantly, using the option can run into
problems if a build system wants to do layering checks on the
generated code. So...

- Document the problem
- Remove the warning

It may make sense to flip the default on this option back in the future and even deprecated/eventually remove it. The other choice would be to figure out how to make the generation work with the
transitive `import public` problem.

PiperOrigin-RevId: 820681829