Account for 8-byte alignment in arena allocation size checks.
The default alignment of arenas is 8 bytes on both 32-bit and 64-bit platforms, but message/field types may not be 8-byte aligned on 32-bit (for example, if they have an odd number of 4-byte fields). If their size happens to not be a multiple of 8, these tests will fail, since `sizeof(T) != Align8(sizeof(T))`.

PiperOrigin-RevId: 822693390