Removed indirection for sub-messages.
To do this, we use weak definitions of actual MiniTables for tree shaking, instead of pointers to MiniTables.  This means that a tree shaken MiniTable will take up `sizeof(upb_MiniTable)` space, instead of `sizeof(void*)`, which is a size regression.  On the other hand, this saves `sizeof(void*)` bytes from every sub-message field, so it may be a size improvement overall.

PiperOrigin-RevId: 823966877