Update Message_index to account for oneof's
This is mostly patterned off of `Message_method_missing`, but that
function does a bunch of extra stuff (to handle, e.g., `has_` and
`clear_` and setter methods), while `Message_index` only needs to handle
getter methods. So instead of copying the `Message_method_missing`
implementation verbatim, we can "constant fold" away the parts of the
method that we know cannot (or should not) happen for `msg[...]` calls.