Removed indirection for sub-messages.
To do this, we use `__attribute__((weakref()))` to weakly reference a placeholder definition for tree shaken sub-messages.  This provides the same tree shaking behavior we had before, but without needing an extra indirection.  If we had discovered this technique before, we would have had no reason to add the indirection in cl/640369522.

This will reduce the code size and memory overhead of any MiniTables with sub-messages in them, because we no longer need an extra `void*` per sub-message field to store the extra indirect pointer.

It also should improve efficiency a bit since there is one fewer indirection when recursing into a sub-message field.

PiperOrigin-RevId: 829022414