Define implicit conversion of `RepeatedPtrOverPtrsIterator` from non-const iterator for const iterators.
When `Element` is already non-const, this implicit conversion is the copy constructor. `-Wdeprecated-copy-with-user-provided-copy` throws an error if copy assignment is used for a type with a user-provided copy constructor and no user-provided copy assignment operator.

Rather than provide a copy assignment operator, we can conditionally disable this conversion when `Element` is not const.

PiperOrigin-RevId: 829436831