Refactor how inline string fields are tracked for donation.
Instead of a separate HasBits object with one bit per field we track donation in one arbitrary bit in the capacity.
When the string IsLong that bit in the capacity determines if it is donated.

We also remove all the plumbing for these donation bits, on demand arena destructor, etc.

PiperOrigin-RevId: 843790247