Introduce RepeatedFieldOrProxy to handle implicit conversion from either legacy repeated field containers or proxies.
We have intentionally made `RepeatedFieldProxy` un-constructible by users, so our implementation can assume it is backed by a repeated field in a message and not a legacy repeated field container. However, to ease migration along, it will be very handy to have a type which can point to either the legacy container or a proxied repeated field. This is meant to be a migration tool more than a permanent solution.

`RepeatedFieldOrProxy` has an identical interface to `RepeatedField`, with the addition of a few implicit conversion constructors from both the legacy containers and from `RepeatedFieldProxy`. We can replace `RepeatedField<>&` (and variants) function parameters with `RepeatedFieldOrProxy<>` without needing to update any of callsites. Once all callsites are using proxies, we may finally switch this parameter out for `RepeatedFieldProxy<>`.

Once the backing of proxied repeated fields changes, `RepeatedFieldOrProxy` will incur a runtime cost over `RepeatedFieldProxy`, as accesses/mutations may need to branch on whether they are backed by a legacy container type or the new backing for proxied fields.

PiperOrigin-RevId: 845784870