Added a test that verifies existing enum behavior for Ruby.
Note that this is a behavior where JRuby "native" has different behavior than CRuby or JRuby FFI.

The CRuby / FFI behavior is a longstanding nonconformance, which is to treat all enums as open. While this is nonconformant, we keep this behavior because it is more useful than respecting closed enum semantics.

PiperOrigin-RevId: 850508608